// packages/temporal-bun-sdk/src/proto/temporal/api/enums/v1/command_type_pb.ts
// @generated by protoc-gen-es v1.10.0 with parameter "target=ts,import_extension=ts"
// @generated from file temporal/api/enums/v1/command_type.proto (package temporal.api.enums.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { proto3 } from '@bufbuild/protobuf'

/**
 * Whenever this list of command types is changed do change the function shouldBufferEvent in mutableStateBuilder.go to make sure to do the correct event ordering.
 *
 * @generated from enum temporal.api.enums.v1.CommandType
 */
export enum CommandType {
  /**
   * @generated from enum value: COMMAND_TYPE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: COMMAND_TYPE_SCHEDULE_ACTIVITY_TASK = 1;
   */
  SCHEDULE_ACTIVITY_TASK = 1,

  /**
   * @generated from enum value: COMMAND_TYPE_REQUEST_CANCEL_ACTIVITY_TASK = 2;
   */
  REQUEST_CANCEL_ACTIVITY_TASK = 2,

  /**
   * @generated from enum value: COMMAND_TYPE_START_TIMER = 3;
   */
  START_TIMER = 3,

  /**
   * @generated from enum value: COMMAND_TYPE_COMPLETE_WORKFLOW_EXECUTION = 4;
   */
  COMPLETE_WORKFLOW_EXECUTION = 4,

  /**
   * @generated from enum value: COMMAND_TYPE_FAIL_WORKFLOW_EXECUTION = 5;
   */
  FAIL_WORKFLOW_EXECUTION = 5,

  /**
   * @generated from enum value: COMMAND_TYPE_CANCEL_TIMER = 6;
   */
  CANCEL_TIMER = 6,

  /**
   * @generated from enum value: COMMAND_TYPE_CANCEL_WORKFLOW_EXECUTION = 7;
   */
  CANCEL_WORKFLOW_EXECUTION = 7,

  /**
   * @generated from enum value: COMMAND_TYPE_REQUEST_CANCEL_EXTERNAL_WORKFLOW_EXECUTION = 8;
   */
  REQUEST_CANCEL_EXTERNAL_WORKFLOW_EXECUTION = 8,

  /**
   * @generated from enum value: COMMAND_TYPE_RECORD_MARKER = 9;
   */
  RECORD_MARKER = 9,

  /**
   * @generated from enum value: COMMAND_TYPE_CONTINUE_AS_NEW_WORKFLOW_EXECUTION = 10;
   */
  CONTINUE_AS_NEW_WORKFLOW_EXECUTION = 10,

  /**
   * @generated from enum value: COMMAND_TYPE_START_CHILD_WORKFLOW_EXECUTION = 11;
   */
  START_CHILD_WORKFLOW_EXECUTION = 11,

  /**
   * @generated from enum value: COMMAND_TYPE_SIGNAL_EXTERNAL_WORKFLOW_EXECUTION = 12;
   */
  SIGNAL_EXTERNAL_WORKFLOW_EXECUTION = 12,

  /**
   * @generated from enum value: COMMAND_TYPE_UPSERT_WORKFLOW_SEARCH_ATTRIBUTES = 13;
   */
  UPSERT_WORKFLOW_SEARCH_ATTRIBUTES = 13,

  /**
   * @generated from enum value: COMMAND_TYPE_PROTOCOL_MESSAGE = 14;
   */
  PROTOCOL_MESSAGE = 14,

  /**
   * @generated from enum value: COMMAND_TYPE_MODIFY_WORKFLOW_PROPERTIES = 16;
   */
  MODIFY_WORKFLOW_PROPERTIES = 16,

  /**
   * @generated from enum value: COMMAND_TYPE_SCHEDULE_NEXUS_OPERATION = 17;
   */
  SCHEDULE_NEXUS_OPERATION = 17,

  /**
   * @generated from enum value: COMMAND_TYPE_REQUEST_CANCEL_NEXUS_OPERATION = 18;
   */
  REQUEST_CANCEL_NEXUS_OPERATION = 18,
}
// Retrieve enum metadata with: proto3.getEnumType(CommandType)
proto3.util.setEnumType(CommandType, 'temporal.api.enums.v1.CommandType', [
  { no: 0, name: 'COMMAND_TYPE_UNSPECIFIED' },
  { no: 1, name: 'COMMAND_TYPE_SCHEDULE_ACTIVITY_TASK' },
  { no: 2, name: 'COMMAND_TYPE_REQUEST_CANCEL_ACTIVITY_TASK' },
  { no: 3, name: 'COMMAND_TYPE_START_TIMER' },
  { no: 4, name: 'COMMAND_TYPE_COMPLETE_WORKFLOW_EXECUTION' },
  { no: 5, name: 'COMMAND_TYPE_FAIL_WORKFLOW_EXECUTION' },
  { no: 6, name: 'COMMAND_TYPE_CANCEL_TIMER' },
  { no: 7, name: 'COMMAND_TYPE_CANCEL_WORKFLOW_EXECUTION' },
  { no: 8, name: 'COMMAND_TYPE_REQUEST_CANCEL_EXTERNAL_WORKFLOW_EXECUTION' },
  { no: 9, name: 'COMMAND_TYPE_RECORD_MARKER' },
  { no: 10, name: 'COMMAND_TYPE_CONTINUE_AS_NEW_WORKFLOW_EXECUTION' },
  { no: 11, name: 'COMMAND_TYPE_START_CHILD_WORKFLOW_EXECUTION' },
  { no: 12, name: 'COMMAND_TYPE_SIGNAL_EXTERNAL_WORKFLOW_EXECUTION' },
  { no: 13, name: 'COMMAND_TYPE_UPSERT_WORKFLOW_SEARCH_ATTRIBUTES' },
  { no: 14, name: 'COMMAND_TYPE_PROTOCOL_MESSAGE' },
  { no: 16, name: 'COMMAND_TYPE_MODIFY_WORKFLOW_PROPERTIES' },
  { no: 17, name: 'COMMAND_TYPE_SCHEDULE_NEXUS_OPERATION' },
  { no: 18, name: 'COMMAND_TYPE_REQUEST_CANCEL_NEXUS_OPERATION' },
])